feat: i have no words

+++ apps/api/lib/repositories/url.repository.ts
@@ -10,15 +10,20 @@ export class UrlRepository {
 		return await this.db.insertInto("shortn.urls").values(values).returning(["id"]).executeTakeFirstOrThrow();
 	}
 
-	async updateUrl(id: number, values: Updateable<ShortnUrls>) {
-		return await this.db.updateTable("shortn.urls").set(values).where("id", "=", id).executeTakeFirstOrThrow();
+	async updateUrl(id: number, userId: number, values: Updateable<ShortnUrls>) {
+		return await this.db
+			.updateTable("shortn.urls")
+			.set(values)
+			.where("id", "=", id)
+			.where("userId", "=", userId)
+			.executeTakeFirstOrThrow();
 	}
 
-	async deleteUrl(id: number) {
-		return await this.db.deleteFrom("shortn.urls").where("id", "=", id).execute();
+	async deleteUrl(id: number, userId: number) {
+		return await this.db.deleteFrom("shortn.urls").where("id", "=", id).where("userId", "=", userId).execute();
 	}
 
-	async findUrl(id: number) {
+	async findUrlById(id: number) {
 		return await this.db.selectFrom("shortn.urls").selectAll().where("id", "=", id).executeTakeFirst();
 	}
 
@@ -26,8 +31,8 @@ export class UrlRepository {
 		return await this.db
 			.selectFrom("shortn.urls")
 			.selectAll()
-			.where("userId", "=", userId)
 			.where("url", "=", url)
+			.where("userId", "=", userId)
 			.executeTakeFirst();
 	}
 
@@ -41,16 +46,27 @@ export class UrlRepository {
 		return await this.db.insertInto("shortn.customUrls").values(values).returning(["id"]).executeTakeFirstOrThrow();
 	}
 
-	async updateCustomUrl(id: number, values: Updateable<ShortnCustomUrls>) {
-		return await this.db.updateTable("shortn.customUrls").set(values).where("id", "=", id).executeTakeFirstOrThrow();
+	async updateCustomUrl(id: number, userId: number, values: Updateable<ShortnCustomUrls>) {
+		return await this.db
+			.updateTable("shortn.customUrls")
+			.set(values)
+			.where("id", "=", id)
+			.where("userId", "=", userId)
+			.returningAll()
+			.executeTakeFirstOrThrow();
 	}
 
-	async deleteCustomUrl(id: number) {
-		return await this.db.deleteFrom("shortn.customUrls").where("id", "=", id).execute();
+	async deleteCustomUrl(id: number, userId: number) {
+		return await this.db.deleteFrom("shortn.customUrls").where("id", "=", id).where("userId", "=", userId).execute();
 	}
 
-	async findCustomUrl(id: number) {
-		return await this.db.selectFrom("shortn.customUrls").selectAll().where("id", "=", id).executeTakeFirst();
+	async findCustomUrlById(id: number, userId: number) {
+		return await this.db
+			.selectFrom("shortn.customUrls")
+			.selectAll()
+			.where("id", "=", id)
+			.where("userId", "=", userId)
+			.executeTakeFirst();
 	}
 
 	async findCustomUrlByCustomCode(customCode: string) {
